+++ src/tp2/CurvarEsquerda.java
@@ -6,7 +6,7 @@ public class CurvarEsquerda extends Mensagem {
 	float angulo;
 	
 	public CurvarEsquerda(int raio, float angulo) {
-		super(TIPO_MENSAGEM.CURVA_ESQ, String.format("Curva à Esquerda (raio %d, ângulo %f) \n", raio, angulo));
+		super(TIPO_MENSAGEM.CURVA_ESQ, String.format("Curva à Esquerda (raio %d, ângulo %d) \n", raio, (int)angulo));
 		this.raio = raio;
 		this.angulo = angulo;
 	}
